+++ docs/releases.md
@@ -3,6 +3,13 @@
 This page will give you the change that are occuring when a new version has been published on pypi.
 The changes have been tracked starting version 0.1.0
 
+## 0.2.1
+
+* requestCreator can now set `setSearch` and `removeSearch`
+* avoiding modifying initial request from `getReport` 
+* fixing report output decryption from `getReport`
+* adding throwing an `TimeoutError` when API Gateway of CJA is not responding.
+
 ## 0.2.0
 
 * adding support for Oauth Server to Server configuration file